pyproject.toml
[tool.poetry]
name = "verses"
version = "0.1.0"
description = "python tools for aws and more"
authors = ["user.name <gmail@hdknr.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
django-environ = "^0.8.1"
boto3 = "^1.20.38"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
